+++ code/prboom/p_doors.c
@@ -508,7 +508,7 @@ int EV_VerticalDoor
 
   // get the sector on the second side of activating linedef
   sec = sides[line->sidenum[1]].sector;
-  secnum = sec-sectors;
+  secnum = (int)(sec-sectors);
 
   /* if door already has a thinker, use it
    * cph 2001/04/05 -
